updates to snp-calling and related tools
