+++ bin/tools.py
@@ -1,7 +1,33 @@
-import os, sys
+#%%
+import os, sys, re
+import pandas as pd, numpy as np
+from Bio import SeqIO
+from Bio.Seq import Seq
 from Bio import Align
-import numpy as np
-import re
+from Bio.SeqRecord import SeqRecord
+
+
+def init_aligner(mode='global', open_gap=-1.0, x_gap=-0.1):
+	aligner = Align.PairwiseAligner()
+	aligner.mode = mode
+	aligner.open_gap_score = open_gap
+	aligner.extend_gap_score = x_gap
+	aligner.target_end_gap_score = 0.0
+	aligner.query_end_gap_score = 0.0
+	return aligner
+
+def get_boundaries(str):
+    gap_prefix = re.compile('^[-]+')
+    gap_suffix = re.compile('[-]+$')
+    # return a tuple giving indices of subsequence without gap prefix and suffix
+    start, stop = 0, len(str)
+    left = gap_prefix.findall(str)
+    right = gap_suffix.findall(str)
+    if left:
+        start = len(left[0])
+    if right:
+        stop = len(str) - len(right[0])
+    return start, stop
 
 def flex_translate(nt_seq, debug=False):
 	"""
@@ -44,27 +70,6 @@ def flex_translate(nt_seq, debug=False):
 	best_seq.description = nt_seq.description
 	return best_seq, best_frame, min_count
 
-def init_aligner(mode='global', open_gap=-1.0, x_gap=-0.1):
-	aligner = Align.PairwiseAligner()
-	aligner.mode = mode
-	aligner.open_gap_score = open_gap
-	aligner.extend_gap_score = x_gap
-	aligner.target_end_gap_score = 0.0
-	aligner.query_end_gap_score = 0.0
-	return aligner
-
-def get_boundaries(str):
-    gap_prefix = re.compile('^[-]+')
-    gap_suffix = re.compile('[-]+$')
-    # return a tuple giving indices of subsequence without gap prefix and suffix
-    res = [0,len(str)]
-    left = gap_prefix.findall(str)
-    right = gap_suffix.findall(str)
-    if left:
-        res[0] = len(left[0])
-    if right:
-        res[1] = len(str) - len(right[0])
-    return res
 
 # performs a pairwise alignment between two Biopython SeqRecord objects
 def pairwise_alignment(aligner, ref, qry):
@@ -83,8 +88,44 @@ def pairwise_alignment(aligner, ref, qry):
 	The aligned reference and query sequences are returned as strings.
 	"""
 	ref_aln, qry_aln = next(aligner.align(str(ref.seq), str(qry.seq)))
-	# cut down the alignment to only the region of interest (that which overlaps with reference)
+
 	start, stop = get_boundaries(ref_aln)
 	ref_aln = ref_aln[start:stop]
 	qry_aln = qry_aln[start:stop]
 	return ref_aln, qry_aln
+
+
+def get_mutations(ref, qry):
+	mutations = []
+	insertion = []
+	deletion = []
+
+	ref_pos = 0
+	for ref_char, qry_char in zip(ref, qry):
+		# increment the ref_pos at any valid reference position
+		if ref_char != '-':  
+			ref_pos += 1
+
+		if ref_char == '-':     # insertion
+			if deletion: 
+				mutations += [("".join(deletion), ref_pos-1, "-")]
+				deletion = []	
+			insertion += qry_char
+		elif qry_char == '-':   # deletion
+			if insertion:
+				mutations += [("-", ref_pos-1, "".join(insertion))]
+				insertion = []
+			deletion += ref_char
+
+		else:					# neither insertion nor deletion
+			if insertion:
+				mutations += [("-", ref_pos-1, "".join(insertion))]
+				insertion = []
+			if deletion: 
+				mutations += [("".join(deletion), ref_pos-1, "-")]
+				deletion = []
+
+			if ref_char != qry_char and qry_char != "X": 	# standard snp mismatch 
+				mutations += [(ref_char, ref_pos, qry_char)]
+
+	return mutations
